Organize tx clients
