+++ nibiru/clients/tx/__init__.py
@@ -1,0 +1,1 @@
+from .tx_client import TxClient  # noqa
