--- nibiru/sdks/__init__.py
@@ -1,14 +1,0 @@
-import abc
-
-import nibiru
-
-
-class PythonMsg(abc.ABC):
-    @abc.abstractmethod
-    def to_pb(self) -> nibiru.ProtobufMessage:
-        """
-        Generate the protobuf message
-
-        Returns:
-            Any: The protobuff mesage
-        """
