--- nibiru/sdks/tx/__init__.py
@@ -1,1 +1,0 @@
-from nibiru.sdks.tx.tx_client import TxClient  # noqa
