+++ tests/pricefeed_test.py
@@ -7,7 +7,7 @@ from nibiru_proto.proto.cosmos.base.abci.v1beta1.abci_pb2 import TxResponse
 
 import nibiru
 import tests
-from nibiru.sdks.tx.pricefeed import MsgPostPrice
+from nibiru.clients.tx.pricefeed import MsgPostPrice
 from tests import dict_keys_must_match, transaction_must_succeed
 
 WHITELISTED_ORACLES: List[str] = [
